Move conversion from full object to separate file
